[Ref]: 영화 장르 데이터 구조 추가 및 MovieDetail 개선
- Genre 구조체 추가
  - Codable 및 Identifiable을 구현하여 API 응답 데이터와 호환
  - 영화 장르 정보를 보다 명확하게 관리할 수 있도록 개선

- MovieDetail 구조체 개선
  - genres 필드를 추가하여 영화의 장르 정보를 포함
  - runtime 필드를 추가하여 영화의 상영 시간 정보 포함
  - backDropURL 계산 프로퍼티 추가하여 배경 이미지 URL 처리

- MovieCategory 구조체 수정
  - MovieCategory.swift를 기존 위치에서 제거하고 Movie.swift 내에 통합
  - MovieListType을 유지하면서 불필요한 카테고리 주석 정리

- SignUpResponse.swift 제거 후 User에 병합
