+++ CineHive/CineHive/Models/Genre.swift
@@ -1,0 +1,13 @@
+//
+//  Genre.swift
+//  CineHive
+//
+//  Created by 이종민 on 3/6/25.
+//
+
+import Foundation
+
+struct Genre: Codable, Identifiable {
+    let id: Int
+    let name: String
+}
