--- CineHive/CineHive/Models/SignUpResponse.swift
@@ -1,13 +1,0 @@
-//
-//  SignUpResponse.swift
-//  CineHive
-//
-//  Created by 존진 on 2/28/25.
-//
-
-import Foundation
-
-struct SignUpResponse: Codable {
-    let message: String
-    let status: String
-}
